Move modal utilities to separate module
